requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
ujson==5.8.0
loguru==0.7.2
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.7
